blog_comments/requirements.txt
asgiref==3.5.0
Django==3.2
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
